refactor: moved index.js to main.js, and index.test.js to main.test.js
